pyproject.toml
[tool.poetry]
name = "ollamachat"
version = "0.1.0"
description = ""
authors = ["Anson <anson.aw0618@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
websocket-client = "==1.9.0"
pillow = "^12.0.0"
fastapi = "==0.119.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
